agents/ForgePilot/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
httpx>=0.27.0
requests>=2.32.0
orjson>=3.10.0
loguru>=0.7.2
rich>=13.7.1
pyyaml>=6.0.1
typing-extensions>=4.12.0

uvloop>=0.19.0; sys_platform != "win32"

pytest>=8.2.0
pytest-asyncio>=0.23.7